Folder location was changed
