+++ src/CodeReview.Evaluator/Services/IIssueFilterFactory.cs
@@ -1,4 +1,5 @@
 ﻿using GodelTech.CodeReview.Evaluator.Models;
+using GodelTech.CodeReview.Evaluator.Services.IssueFilters;
 
 namespace GodelTech.CodeReview.Evaluator.Services
 {
