+++ src/CodeReview.Evaluator/Services/ILocDetailsFilterFactory.cs
@@ -1,4 +1,5 @@
 ﻿using GodelTech.CodeReview.Evaluator.Models;
+using GodelTech.CodeReview.Evaluator.Services.LocDetailsFilters;
 
 namespace GodelTech.CodeReview.Evaluator.Services
 {
