+++ src/CodeReview.Evaluator/Services/IssueFilterFactory.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
 using GodelTech.CodeReview.Evaluator.Models;
+using GodelTech.CodeReview.Evaluator.Services.IssueFilters;
 
 namespace GodelTech.CodeReview.Evaluator.Services
 {
