+++ src/CodeReview.Evaluator/Services/IssueFilters/IIssueFilter.cs
@@ -1,6 +1,6 @@
 ﻿using GodelTech.CodeReview.Evaluator.Models;
 
-namespace GodelTech.CodeReview.Evaluator.Services
+namespace GodelTech.CodeReview.Evaluator.Services.IssueFilters
 {
     public interface IIssueFilter
     {
