+++ src/CodeReview.Evaluator/Services/LocDetailsFilters/CompositeFilter.cs
@@ -2,14 +2,14 @@
 using System.Linq;
 using GodelTech.CodeReview.Evaluator.Models;
 
-namespace GodelTech.CodeReview.Evaluator.Services
+namespace GodelTech.CodeReview.Evaluator.Services.LocDetailsFilters
 {
-    public class CompositeLocDetailsFilter : ILocDetailsFilter
+    public class CompositeFilter : ILocDetailsFilter
     {
         private readonly bool _trueOnAnyMatch;
         private readonly ILocDetailsFilter[] _filters;
 
-        public CompositeLocDetailsFilter(bool trueOnAnyMatch, params ILocDetailsFilter[] filters)
+        public CompositeFilter(bool trueOnAnyMatch, params ILocDetailsFilter[] filters)
         {
             _trueOnAnyMatch = trueOnAnyMatch;
             _filters = filters ?? throw new ArgumentNullException(nameof(filters));
