+++ src/CodeReview.Evaluator/Services/LocDetailsFilters/GenericFilter.cs
@@ -1,20 +1,20 @@
 ﻿using System;
 using GodelTech.CodeReview.Evaluator.Models;
 
-namespace GodelTech.CodeReview.Evaluator.Services
+namespace GodelTech.CodeReview.Evaluator.Services.LocDetailsFilters
 {
-    public class GenericIssueFilter<T> : IIssueFilter
+    public class GenericFilter<T> : ILocDetailsFilter
     {
-        private readonly Func<Issue, T> _valueSelector;
+        private readonly Func<FileLocDetails, T> _valueSelector;
         private readonly Func<T, bool> _predicate;
 
-        public GenericIssueFilter(Func<Issue, T> valueSelector, Func<T, bool> predicate)
+        public GenericFilter(Func<FileLocDetails, T> valueSelector, Func<T, bool> predicate)
         {
             _valueSelector = valueSelector ?? throw new ArgumentNullException(nameof(valueSelector));
             _predicate = predicate ?? throw new ArgumentNullException(nameof(predicate));
         }
 
-        public bool IsMatch(Issue issue)
+        public bool IsMatch(FileLocDetails issue)
         {
             if (issue == null) throw new ArgumentNullException(nameof(issue));
